Robustness for go list
Some benchmarks were throwing a "no matches found for regexp" error. This is returned by go list when no user code (package) allocations made it into the memprofile.

This is not actually an error for our purposes. I believe it's simply a correct outcome of no allocations. So now we catch that error and look for that string.

Additionally, adjusted added -memprofilerate=1024*64 to the memprofile creation. As I understand it, this describes resolution/precision of allocations. "1024*64" indicates "only include allocs over 1024*64 bytes". I believe the default is 512000, 512K. So now we might detect more allocations. Expect a small perf hit.

memprofilerate was adjusted in an attempt to fix the error above. It did not fix the error, but seems desirable anyway.