Simplify runAllBenchmarks (#3)
* Make runAllBenchmarks simple and synchronous

Remove the complexity of the semaphore. Kludgy both in implementation and UX. Concurrency will likely be desirable in the future, but let's get back to basics to start.

* Remove hasBeenRun

It does not help any functionality, since that state is handled via getChildren and reveal. It was only assigned, but not read.

* Add concurrency of 2 on runAllBenchmarks

Getting the concurrency working turned out to be finicky. tl;dr stuff about Thenable vs Promise on treeView.reveal. Copilot figured it out, Cursor struggled.