Use gopls to discover benchmarks (#5)
Specifically, use `vscode.executeWorkspaceSymbolProvider` to query benchmarks, instead of `go list` for each package. Way faster.

It’s still kinda “stringy” in the sense that we don’t have a static guarantee that it’s a legit Go benchmark — we check that it’s a func, that it uses the Go naming convention, and is in a test file. We don’t check the full signature, as that would require the AST.

If something meets those criteria and is not a benchmark, it’ll fail to run anyway.