Remove the gcflags from memprofile
I am not convinced those gcflags are helpful, and they might actually be a net negative.

The theory was that inlining or optimizations might hide, or mislead on, the source of allocations.

But! I assume the memprofile is aware of this, and in any case, inlined and optimized code is what the user will run, so let's analyze that.